Moved API logic to a new file
